refactor: modularize the client implementations
OpenSubsonicExtension.kt now delegates the implementation of the Client classes to the ClientImpl classes in the clients/ package.
OpenSubsonicApi.kt has been removed as it was just a mapping between the Client functions and their implementations.
A service/ package has been created containing the previous collections of functions RequestApi.kt and GenreApi.kt and the session/ subpackage containing the shared UserData and Settings objects.
The UserData and Server, now renamed to ServerData classes have been moved to the model/ package
Utils.kt functions have been moved in RequestService.kt as they were related to OkHttp objects
