--- ext/src/main/java/dev/brahmkshatriya/echo/extension/api/OpenSubsonicApi.kt
@@ -1,80 +1,0 @@
-package dev.brahmkshatriya.echo.extension.api
-
-import dev.brahmkshatriya.echo.common.models.Album
-import dev.brahmkshatriya.echo.common.models.Artist
-import dev.brahmkshatriya.echo.common.models.Feed
-import dev.brahmkshatriya.echo.common.models.Shelf
-import dev.brahmkshatriya.echo.common.models.Streamable
-import dev.brahmkshatriya.echo.common.models.Track
-import dev.brahmkshatriya.echo.common.models.User
-import dev.brahmkshatriya.echo.extension.api.album.createAlbumFeed
-import dev.brahmkshatriya.echo.extension.api.album.getAlbum
-import dev.brahmkshatriya.echo.extension.api.album.getTracks
-import dev.brahmkshatriya.echo.extension.api.artist.createArtistFeed
-import dev.brahmkshatriya.echo.extension.api.artist.getArtist
-import dev.brahmkshatriya.echo.extension.api.login.getUserSession
-import dev.brahmkshatriya.echo.extension.api.login.keyLogin
-import dev.brahmkshatriya.echo.extension.api.login.passwordLogin
-import dev.brahmkshatriya.echo.extension.api.login.setUserSession
-import dev.brahmkshatriya.echo.extension.api.track.createTrackFeed
-import dev.brahmkshatriya.echo.extension.api.track.getStreamableMedia
-import dev.brahmkshatriya.echo.extension.api.track.getTrack
-
-class OpenSubsonicApi {
-
-    // Login
-
-    suspend fun onPasswordLogin(data: Map<String, String?>): List<User> {
-        return passwordLogin(data)
-    }
-
-    suspend fun onKeyLogin(data: Map<String, String?>): List<User> {
-        return keyLogin(data)
-    }
-
-    fun setUser(user: User?) {
-        setUserSession(user)
-    }
-
-    fun getUser(): User? {
-        return getUserSession()
-    }
-
-    // Track
-
-    fun loadTrack(track: Track): Track {
-        return getTrack(track)
-    }
-
-    fun loadStreamableMedia(streamable: Streamable, isDownload: Boolean): Streamable.Media {
-        return getStreamableMedia(streamable, isDownload)
-    }
-
-    suspend fun loadTrackFeed(track: Track): Feed<Shelf> {
-        return createTrackFeed(track)
-    }
-
-    // Artist
-
-    suspend fun loadArtist(artist: Artist): Artist {
-        return getArtist(artist)
-    }
-
-    suspend fun loadArtistFeed(artist: Artist): Feed<Shelf> {
-        return createArtistFeed(artist)
-    }
-
-    // Album
-
-    suspend fun loadAlbum(album: Album): Album {
-        return getAlbum(album)
-    }
-
-    suspend fun loadTracks(album: Album): Feed<Track>? {
-        return getTracks(album)
-    }
-
-    suspend fun loadAlbumFeed(album: Album): Feed<Shelf>? {
-        return createAlbumFeed(album)
-    }
-}
