--- ext/src/main/java/dev/brahmkshatriya/echo/extension/api/login/UserSession.kt
@@ -1,6 +1,0 @@
-package dev.brahmkshatriya.echo.extension.api.login
-
-object UserSession {
-    @Volatile
-    var current: UserData = UserData.EMPTY
-}
