+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/dto/endpoints/GetSongsByGenreDto.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class GetSongsByGenreDto(
     @SerialName("subsonic-response")
-    val subsonicResponse: SubsonicResponseDto
+    val subsonicResponse: SubsonicResponseDto,
 ) {
     @Serializable
     data class SubsonicResponseDto(
