+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/dto/types/AlbumDto.kt
@@ -3,8 +3,8 @@ package dev.brahmkshatriya.echo.extension.dto.types
 import dev.brahmkshatriya.echo.common.models.Album
 import dev.brahmkshatriya.echo.common.models.Date
 import dev.brahmkshatriya.echo.common.models.ImageHolder
-import dev.brahmkshatriya.echo.extension.api.request.authenticatedRequest
-import dev.brahmkshatriya.echo.extension.toNetworkRequest
+import dev.brahmkshatriya.echo.extension.service.request.RequestService.authenticatedRequest
+import dev.brahmkshatriya.echo.extension.service.request.RequestService.toNetworkRequest
 import kotlinx.serialization.Serializable
 
 @Serializable
