+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/dto/types/ArtistDto.kt
@@ -2,8 +2,8 @@ package dev.brahmkshatriya.echo.extension.dto.types
 
 import dev.brahmkshatriya.echo.common.models.Artist
 import dev.brahmkshatriya.echo.common.models.ImageHolder
-import dev.brahmkshatriya.echo.extension.api.request.authenticatedRequest
-import dev.brahmkshatriya.echo.extension.toNetworkRequest
+import dev.brahmkshatriya.echo.extension.service.request.RequestService.authenticatedRequest
+import dev.brahmkshatriya.echo.extension.service.request.RequestService.toNetworkRequest
 import kotlinx.serialization.Serializable
 
 @Serializable
