+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/dto/types/GenreDto.kt
@@ -3,7 +3,7 @@ package dev.brahmkshatriya.echo.extension.dto.types
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GenreDto (
+data class GenreDto(
     val value: String,
     val albumCount: Int,
     val songCount: Int,
