+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/dto/types/IndexDto.kt
@@ -3,7 +3,7 @@ package dev.brahmkshatriya.echo.extension.dto.types
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class IndexDto (
+data class IndexDto(
     val name: String,
     val artist: List<ArtistDto>? = null,
 )
