+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/dto/types/RecordLabelDto.kt
@@ -3,6 +3,6 @@ package dev.brahmkshatriya.echo.extension.dto.types
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RecordLabelDto (
+data class RecordLabelDto(
     val name: String,
 )
