+++ ext/src/main/java/dev/brahmkshatriya/echo/extension/service/session/UserSession.kt
@@ -1,0 +1,8 @@
+package dev.brahmkshatriya.echo.extension.service.session
+
+import dev.brahmkshatriya.echo.extension.models.UserData
+
+object UserSession {
+    @Volatile
+    var currentUser: UserData = UserData.EMPTY
+}
